demo-python-service/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
py_eureka_client==0.9.9
flasgger==0.9.7

# Resilience Patterns
circuitbreaker==1.4.0
retrying==1.3.4
tenacity==8.2.3

# Caching & Storage
redis==5.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
cachetools==5.3.2

# Messaging
kafka-python==2.0.2
pika==1.3.2
celery==5.3.4

# Async & Reactive
aiohttp>=3.12.14
aioredis==2.0.1
rx==4.0.4
asyncio==3.4.3

# Monitoring & Tracing
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
structlog==23.2.0

# Security
PyJWT==2.8.0
cryptography==41.0.7
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0

# Database & ORM
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-RESTful==0.3.10
marshmallow==3.20.1

# Utilities
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0
click==8.1.7
gunicorn==21.2.0
APScheduler==3.10.4
uuid==1.30
dataclasses-json==0.6.3
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-flask==1.3.0
pytest-mock==3.12.0
